Separate bundling of parsers
