+++ .gitignore
@@ -3,3 +3,5 @@ npm-debug.log
 /errors
 /test.js
 /.vscode
+
+/dist
