+++ src/parser.js
@@ -1,21 +1,14 @@
 "use strict";
 
-function createError(message, line, column) {
-  // Construct an error similar to the ones thrown by Babylon.
-  const error = new SyntaxError(message + " (" + line + ":" + column + ")");
-  error.loc = { line, column };
-  return error;
-}
-
 function parse(text, opts) {
   let parseFunction;
 
   if (opts.parser === "flow") {
-    parseFunction = parseWithFlow;
+    parseFunction = require("./parsers/flow-parser");
   } else if (opts.parser === "typescript") {
-    parseFunction = parseWithTypeScript;
+    parseFunction = require("./parsers/typescript-parser");
   } else {
-    parseFunction = parseWithBabylon;
+    parseFunction = require("./parsers/babylon-parser");
   }
 
   try {
@@ -33,109 +26,6 @@ function parse(text, opts) {
 
     throw error;
   }
-}
-
-function parseWithFlow(text) {
-  // Inline the require to avoid loading all the JS if we don't use it
-  const flowParser = require("flow-parser");
-
-  const ast = flowParser.parse(text, {
-    esproposal_class_instance_fields: true,
-    esproposal_class_static_fields: true,
-    esproposal_export_star_as: true
-  });
-
-  if (ast.errors.length > 0) {
-    throw createError(
-      ast.errors[0].message,
-      ast.errors[0].loc.start.line,
-      ast.errors[0].loc.start.column
-    );
-  }
-
-  return ast;
-}
-
-function parseWithBabylon(text) {
-  // Inline the require to avoid loading all the JS if we don't use it
-  const babylon = require("babylon");
-
-  const babylonOptions = {
-    sourceType: "module",
-    allowImportExportEverywhere: false,
-    allowReturnOutsideFunction: true,
-    plugins: [
-      "jsx",
-      "flow",
-      "doExpressions",
-      "objectRestSpread",
-      "decorators",
-      "classProperties",
-      "exportExtensions",
-      "asyncGenerators",
-      "functionBind",
-      "functionSent",
-      "dynamicImport"
-    ]
-  };
-
-  try {
-    return babylon.parse(text, babylonOptions);
-  } catch (originalError) {
-    try {
-      return babylon.parse(
-        text,
-        Object.assign({}, babylonOptions, { strictMode: false })
-      );
-    } catch (nonStrictError) {
-      throw originalError;
-    }
-  }
-}
-
-function parseWithTypeScript(text) {
-  const jsx = isProbablyJsx(text);
-  try {
-    try {
-      // Try passing with our best guess first.
-      return tryParseTypeScript(text, jsx);
-    } catch (e) {
-      // But if we get it wrong, try the opposite.
-      return tryParseTypeScript(text, !jsx);
-    }
-  } catch (e) {
-    throw createError(e.message, e.lineNumber, e.column);
-  }
-}
-
-function tryParseTypeScript(text, jsx) {
-  // While we are working on typescript, we are putting it in devDependencies
-  // so it shouldn't be picked up by static analysis
-  const r = require;
-  const parser = r("typescript-eslint-parser");
-  return parser.parse(text, {
-    loc: true,
-    range: true,
-    tokens: true,
-    comment: true,
-    useJSXTextNode: true,
-    ecmaFeatures: { jsx }
-  });
-}
-
-/**
- * Use a naive regular expression until we address
- * https://github.com/prettier/prettier/issues/1538
- */
-function isProbablyJsx(text) {
-  return new RegExp(
-    [
-      "(^[^\"'`]*</)", // Contains "</" when probably not in a string
-      "|",
-      "(^[^/]{2}.*\/>)" // Contains "/>" on line not starting with "//"
-    ].join(""),
-    "m"
-  ).test(text);
 }
 
 module.exports = { parse };
